pyproject.toml
[project]
name = "clingen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.86",
    "chromadb>=1.3.5",
    "cyvcf2>=0.31.4",
    "fastapi>=0.124.0",
    "langchain>=1.1.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.1.0",
    "langchain-ollama>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
